+++ src/main/java/searchengine/repository/LemmaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import searchengine.model.LemmaEntity;
 
 @Repository
-public interface LemmaRepository extends JpaRepository<LemmaEntity, Integer> {
+public interface LemmaRepository extends JpaRepository<LemmaEntity, Long> {
 }
