+++ src/main/java/searchengine/repository/PageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import searchengine.model.PageEntity;
 
 @Repository
-public interface PageRepository extends JpaRepository<PageEntity,Integer> {
+public interface PageRepository extends JpaRepository<PageEntity,Long> {
 }
